+++ core/components/customrequest/lexicon/en/default.inc.php
@@ -6,12 +6,9 @@
  * @subpackage lexicon
  */
 $_lang['customrequest'] = 'CustomRequest';
-$_lang['customrequest_desc'] = 'Use friendly URLs everywhere';
 
 $_lang['customrequest.menu_home'] = 'CustomRequest';
 $_lang['customrequest.menu_home_desc'] = 'Use friendly URLs everywhere';
-
-$_lang['customrequest.debug_mode'] = 'Debug Mode';
 
 $_lang['customrequest.configs'] = 'Configurations';
 $_lang['customrequest.configs_alias'] = 'Alias Path';
@@ -33,12 +30,14 @@ $_lang['customrequest.configs_update'] = 'Update Configuration';
 $_lang['customrequest.configs_urlparams'] = 'URI Parameter';
 $_lang['customrequest.configs_urlparams_desc'] = 'The request/get/post parameter keys, the divided second part of the not found URI are assigned to. If the Regular Expression field not set, the second part is divided at the URI separators \'/\'';
 
-$_lang['customrequest.error_msg'] = 'Error: [[+msg]]';
+$_lang['customrequest.settings'] = '<i class="icon icon-cog"></i>';
+$_lang['customrequest.settings_desc'] = 'Edit the settings of Crosslinks. You can edit the value of a system setting by double-clicking on the \'Value\' table cell or by right-clicking in the table cell.';
 
-$_lang['customrequest.configs_err_invalid'] = 'Invalid config.';
-$_lang['customrequest.configs_err_nf'] = 'Config not found.';
-$_lang['customrequest.configs_err_ns'] = 'Config not specified.';
+$_lang['customrequest.debug_mode'] = 'Debug Mode';
+
+$_lang['customrequest.systemsetting_key_err_nv'] = 'You could only edit settings with the prefix customrequest.';
+$_lang['customrequest.systemsetting_usergroup_err_nv'] = 'Only users with a settings permission or a settings_customrequest permission are allowed to change settings.';
+
 $_lang['customrequest.configs_err_ns_alias_resourceid'] = 'Please fill the alias and/or select a resource.';
 $_lang['customrequest.configs_err_nv_alias_regex'] = 'If no resource is selected, the alias has to be a valid regular expression containing delimiters.';
 $_lang['customrequest.configs_err_nv_regex'] = 'The regular expression has to contain delimiters and it to be valid.';
-$_lang['customrequest.configs_err_save'] = 'An error occurred while trying to save the config.';